LPS-160963 Move Action, Render, and ResourceURLBuilder to portal-kernel
